+++ app/src/main/java/com/rave/simplemath/viewmodel/ProductViewModel.kt
@@ -29,7 +29,7 @@ class ProductViewModel : ViewModel() {
         _product.update { it.copy(isLoading = true) }
         val expression = "$firstArg*$secondArg"
         val productResponse = MathRepo.evaluateExpression(expression)
-        _product.update { it.copy(isLoading = false, product = productResponse) }
+        _product.update { it.copy(isLoading = false, product = productResponse.toInt()) }
     }
 }
